tests/backend/requirements.txt
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-mock==3.11.1
fakeredis==2.18.1
psycopg2-binary==2.9.7
redis==4.6.0
requests==2.31.0
